+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2025-09-20 02:44:41
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: b1ab2726e5356f3c71bd825749973622df945516
+
+## [ForwardWidgets-2kuai] - 2025-09-20 02:44:41
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: e1fb8e276abfd0ad245433e81e99e0f4ab5150cc
+
+## [ForwardWidgets-pack1r] - 2025-09-20 02:44:41
+- 仓库: https://github.com/pack1r/ForwardWidgets
+- 分支: main
+- 新增: f88f510b78bcb03e7b9f819669267b9c3c10768b
+
+
 ## [ForwardWidgets-huangxd] - 2025-09-19 02:54:01
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main